+++ CHANGELOG.md
@@ -9,5 +9,6 @@ This changelog only contains the changes that are unreleased. For changes for in
 
 ### Fixes
 - Remove caching of download urls from CurseForge as per ToS, also remove some others from json
+- Don't export CurseForge mods urls in Modrinth pack exports due to CurseForge ToS, use overides
 
 ### Misc